+++ quotient_of_2_numbers.py
@@ -1,6 +1,6 @@
 #input of 2 numbers
-dividend = int(input("Enter the dividend: "))
-divisor = int(input("Enter the divisor: "))
+dividend = float(input("Enter the dividend: "))
+divisor = float(input("Enter the divisor: "))
 
 #showing the quotient of the numbers
 print(dividend / divisor)
